Error message fixes
use write_stderr(), because the crash dump handler can be called
long before we have set up the elog system. We are also running
in a very limited environment when dealing with the crash, so avoid
anything that could potentially lead to things like memory allocs.

Change error messages to more follow the standard style