wal_decoding: mergme: Finally don't require all wal from the initiation of a logical slot to be present
Instead increase the 'restart_decoding' point everytime we have the knowledge
to do so. For that we track the last serialized snapshot before the beginning
of a transaction. As we have a list of transaction thats in LSN order we can
then efficiently query the oldest transaction's restart_decoding_lsn and set
that as the global restart point.

We have to delay the setting of that restart point up to the point where the
client acknowledges reception of the lsn that triggered updating the restart
point.